-refactored the health module (#33)
-modified some data structure to cope with Firebase integration

+++ app/src/main/java/com/neverlost/ubc/neverlost/objects/Caretaker.java
@@ -6,7 +6,7 @@ package com.neverlost.ubc.neverlost.objects;
 
 public class Caretaker extends User {
 
-    public Caretaker(String name, int age, double weight, double height) {
+    public Caretaker(String name, long age, long weight, long height) {
         super(name, age, weight, height);
     }
